op: cosolidate all op functions into op_fns.c
Now each op functions are quite simple and share a lot of commonalities,
it is easier to maintain moving them into a single source file.
