--- src/mpi/coll/op/opmax.c
@@ -1,28 +1,0 @@
-/*
- * Copyright (C) by Argonne National Laboratory
- *     See COPYRIGHT in top-level directory
- */
-
-#include "mpiimpl.h"
-#include "mpir_op_util.h"
-
-/*
- * In MPI-2.1, this operation is valid only for C integer, Fortran integer,
- * and floating point types (5.9.2 Predefined reduce operations)
- */
-
-void MPIR_MAXF(void *invec, void *inoutvec, MPI_Aint * Len, MPI_Datatype * type)
-{
-    MPI_Aint i, len = *Len;
-
-    switch (MPIR_DATATYPE_GET_RAW_INTERNAL(*type)) {
-#undef MPIR_OP_TYPE_MACRO
-#define MPIR_OP_TYPE_MACRO(mpi_type_, c_type_) MPIR_OP_TYPE_REDUCE_CASE(mpi_type_, c_type_, MPL_MAX)
-            /* no semicolons by necessity */
-            MPIR_OP_TYPE_GROUP(INTEGER)
-                MPIR_OP_TYPE_GROUP(FLOATING_POINT)
-        default:
-            MPIR_Assert(0);
-            break;
-    }
-}
